+++ NEWS.md
@@ -1,3 +1,45 @@
+# wdpar 1.3.1
+
+- CRAN release.
+- Remove LazyData from DESCRIPTION since it is not used.
+- Remove tools R package from DESCRIPTION since it is not used.
+
+# wdpar 1.3.0.2
+
+- Update `wdpa_fetch` function to conform with CRAN policies. Specifically,
+  data are now always saved in a temporary directory by default.
+- Update code in README for downloading developmental version from GitHub.
+- Update internal functions to be more compatible with recent version of the sf
+  R package.
+
+# wdpar 1.3.0.1
+
+- Update `wdpa_url` function to be compatible with changes on
+  <https://www.protectedplanet.net/en>.
+- Update `wdpa_read` function to be compatible with new global dataset format.
+- Update `st_erase_overlaps` function to be more resilient against topology
+  errors (#33).
+
+# wdpar 1.3.0.0
+
+- Fix URLs for CRAN.
+- Increase test coverage.
+
+# wdpar 1.2.0.0
+
+- Update `wdpa_clean` function with new `retain_status` parameter to specify
+  which protected areas should be retained during the cleaning process (#33).
+  Defaults to a `character` vector containing `"Designated"`, `"Inscribed"`,
+  `"Established"` to indicate that protected areas with these status
+  (i.e. per `"STATUS"` column) are retained. Thus the default behavior from
+  previous versions remains unchanged.
+
+# wdpar 1.1.0.0
+
+- Update `wdpa_clean` with new `exclude_unesco` parameter to specify if
+  UNESCO Biosphere Reserves should be included (#33). Defaults to `TRUE` such
+  that default behavior remains unchanged from previous versions.
+
 # wdpar 1.0.6
 
 - CRAN release.